[sailfish-connman-iptables-plugin] Restructuring of code for unit test needs.
