+++ src/sailfish-iptables.c
@@ -46,662 +46,18 @@
 #include <sys/socket.h>
 #include <netdb.h>
 #include <stdlib.h>
+#include <dbus/dbus.h>
+#include <glib.h>
 
 #include "sailfish-iptables.h"
+#include "sailfish-iptables-dbus.h"
+#include "sailfish-iptables-validate.h"
+#include "sailfish-iptables-parameters.h"
+#include "sailfish-iptables-utils.h"
 
 #define ERR(fmt,arg...) connman_error(fmt, ## arg)
 #define DBG(fmt,arg...) connman_debug(fmt, ## arg)
 
-const gchar const * OP_STR[] = {"Add", "Remove", "Undefined"};
-const gchar const * RESULT_STR[] = {
-	"Ok",
-	"Invalid IP",
-	"Invalid port",
-	"Invalid port range",
-	"Invalid service name",
-	"Invalid protocol",
-	"Invalid policy",
-	"Invalid file path",
-	"Cannot process rule",
-	"Cannot perform operation",
-};
-const gchar * EMPTY_STR = "";
-
-static gboolean negated_ip_address(const gchar* ip)
-{
-	return ip && ip[0] == '!';
-}
-
-gchar* get_protocol_for_service(const gchar *service)
-{
-	if(service && *service)
-	{
-		struct servent *s = getservbyname(service, NULL);
-		if(s)
-			return g_strdup(s->s_proto);
-	}
-	return NULL;
-}
-
-gchar* get_protocol_for_port(guint16 port)
-{
-	if(port)
-	{
-		struct servent *s = getservbyport(htons(port), NULL);
-		if(s)
-			return g_strdup(s->s_proto);	
-	}
-	return NULL;
-}
-
-static gboolean validate_address(gint type, const gchar* address)
-{
-	struct addrinfo hints;
-	struct addrinfo *result = NULL;
-	
-	memset(&hints, 0, sizeof(struct addrinfo));
-	
-	hints.ai_family = (type == IPV6 ? AF_INET6 : AF_INET);
-	hints.ai_flags = AI_NUMERICHOST;
-	
-	// Success
-	if(!getaddrinfo(address,NULL,&hints,&result))
-	{
-		if(result)
-			freeaddrinfo(result);
-		return true;
-	}
-	return false;
-}
-
-static guint32 mask_to_cidr(gint type, const gchar* mask_address)
-{
-	gint index = 0, b = 0;
-	guint32 mask = 0, bits = 0;
-	gint i = (type == IPV6 ? IPV6_MASK_MAX : IPV4_MASK_MAX);	
-	
-	gchar **ip_tokens = NULL;
-	
-	if(!mask_address)
-		return G_MAXUINT32;
-	
-	ip_tokens = g_strsplit(mask_address,IP_DELIM,4);
-	
-	if(ip_tokens)
-	{
-		// Mask was given as cidr
-		if(g_strv_length(ip_tokens) == 1)
-		{
-			mask = ((guint32)g_ascii_strtoull(ip_tokens[0],NULL,10));
-		}
-		// Dot notation
-		else if(g_strv_length(ip_tokens) == 4)
-		{
-			for(index = 0; index < 4 && ip_tokens[index]; index++)
-			{
-				b = 24 - 8 * index; // 24,16,8,0
-				mask += ((guint32)g_ascii_strtoull(ip_tokens[index],NULL,10)) << b;
-			}
-		}
-	}
-	
-	g_strfreev(ip_tokens);
-	
-	// Return protocol mask max (32/128)
-	if (mask == G_MAXUINT32)
-		return i;
-
-	// Between 0 and protocol mask max, return given mask
-	else if(mask <= i)
-		return mask;
-
-	// Value between protocol max and 2^32, calculate cidr mask
-	// Create cidr notation (bitmask for nw mask)
-	bits = G_MAXUINT32 - 1;
-	while(--i >= 0 && mask != bits)
-		bits <<= 1;
-
-	return i;
-}
-
-gchar* format_ip(gint type, const gchar* ip)
-{	
-	gchar *formatted_ip = NULL;
-	gchar **ip_and_mask = NULL;
-	gint mask_max = 0;
-	
-	if(!ip)
-		return NULL;
-		
-	mask_max = (type == IPV6 ? IPV6_MASK_MAX : IPV4_MASK_MAX);
-		
-	ip_and_mask = g_strsplit(ip,IP_MASK_DELIM,2);
-
-	if(ip_and_mask && g_strv_length(ip_and_mask) == 2)
-	{
-		guint32 mask = mask_to_cidr(type, ip_and_mask[1]);
-	
-		/* 	TODO: when a IP (not network) is given with a bitmask iptables
-			command changes the IP address to network, thus this cannot be
-			removed using the API as such rule does not exist in iptables -
-			the new one with network does.
-		*/
-		
-		// Proper mask, between 0 and 32/128
-		if(mask && mask < mask_max)
-			formatted_ip = g_strdup_printf("%s/%u",
-				ip_and_mask[0], mask);
-
-		/* 	Iptables command removes /32 (or /128 IPv6) from the end, we do the
-			same, also for 0, TODO: if 0 given iptables sets 0.0.0.0/0 (any)
-		*/
-		else if((mask && mask == mask_max) || !mask)
-			formatted_ip = g_strdup_printf("%s", ip_and_mask[0]);
-		// TODO: this may not be reached
-		else
-			formatted_ip = g_strdup_printf("%s/%s",
-				ip_and_mask[0], ip_and_mask[1]);
-	}
-	// No mask separator found
-	else
-		formatted_ip = g_strdup(ip);
-		
-	DBG("Formatted IP: %s",formatted_ip);
-
-	g_strfreev(ip_and_mask);
-	
-	return formatted_ip;
-	
-}
-
-static gboolean validate_ip_mask(gint type, const gchar* mask)
-{
-	if(type == IPV4 && mask)
-	{
-		// Dot notation mask
-		if(strchr(mask,'.'))
-			return validate_address(type, mask);
-
-		guint32 int_mask = g_ascii_strtoull(mask,NULL,10);
-
-		// 0, acceptable
-		if(!int_mask || 
-			(int_mask & (~int_mask >> 1)) ||
-			int_mask == G_MAXUINT32)
-			return true;
-	}
-	
-	return false;
-}
-
-static gboolean validate_ip_address(gint type, const gchar* ip)
-{
-	gboolean rval = false;
-	
-	if(ip && *ip)
-	{
-		const gchar *address = NULL;
-		gchar **ip_and_mask = NULL;
-		
-		// Allow negation
-		if(negated_ip_address(ip))
-			address = &ip[1];
-		else
-			address = ip;
-			
-		// Check for mask
-		ip_and_mask = g_strsplit(address,IP_MASK_DELIM,2);
-		
-		// Both IP and mask are defined
-		if(ip_and_mask && g_strv_length(ip_and_mask) == 2)
-		 	rval = validate_address(type,ip_and_mask[0]) &&
-		 			validate_ip_mask(type, ip_and_mask[1]);
-		else
-			rval = address && validate_address(type, address);
-		
-		g_strfreev(ip_and_mask);
-	}
-	return rval;
-}
-
-static guint16 validate_service_name(const gchar *service)
-{
-	if(service && *service)
-	{	
-		struct servent *s = getservbyname(service, NULL);
-		if(s)
-			return ntohs(s->s_port);	
-	}
-	return 0;
-}
-
-gboolean validate_protocol(const gchar *protocol)
-{
-	if(protocol && *protocol)
-	{
-		struct protoent *p = getprotobyname(protocol);
-		if(p)
-			return true;
-	}
-	return false;
-}
-
-static gboolean validate_port(guint16 port)
-{
-	return port && port < 0xFFFF;
-}
-
-static rule_operation validate_operation(const gchar *operation)
-{
-	rule_operation op = UNDEFINED;
-	
-	if(operation && *operation)
-	{
-		if(!g_ascii_strcasecmp(operation,"ADD"))
-			op = ADD;
-
-		if(!g_ascii_strcasecmp(operation,"REMOVE"))
-			op = REMOVE;
-	}
-	
-	return op;
-}
-
-static gboolean validate_path(const gchar *path)
-{
-	if(path && *path)
-	{
-		// Do proper validation in connman, or add our own validation rules here
-		return true;
-	}
-	return false;
-}
-
-static gboolean validate_policy(const gchar* policy)
-{
-	if(policy && *policy)
-	{
-		if(!g_strcmp0(policy,IPTABLES_ACCEPT) ||
-			!g_strcmp0(policy,IPTABLES_DROP))
-			return true;
-	}
-	return false;
-}
-
-gchar** get_port_range_tokens(const gchar* port_str)
-{
-	if(port_str && *port_str)
-		return g_strsplit(port_str,PORT_RANGE_DELIM,2);
-	return NULL;
-}
-
-gchar *port_to_str(rule_params *params)
-{
-	gchar* port_str = NULL;
-	
-	if(params)
-	{
-		if(params->args == ARGS_PORT_RANGE || 
-			params->args == ARGS_IP_PORT_RANGE)
-			port_str = g_strdup_printf("%u:%u",params->port[0],params->port[1]);
-		else
-			port_str = g_strdup_printf("%u",params->port[0]);
-	}
-	return port_str;
-}
-
-const char* api_result_message(api_result result)
-{
-	if(result >= OK && result <= INVALID)
-		return RESULT_STR[result];
-		
-	return "";
-}
-
-void rule_params_free(rule_params *params)
-{
-	if(params)
-	{
-		g_free(params->ip);
-		g_free(params->service);
-		g_free(params->protocol);
-		g_free(params->path);
-		g_free(params->table);
-		g_free(params->policy);
-		g_free(params);
-	}
-}
-
-rule_params* rule_params_new(rule_args args)
-{
-	rule_params *params = g_new0(rule_params,1);
-	params->ip = NULL;
-	params->ip_negate = false;
-	params->service = NULL;
-	params->port[0] = params->port[1] = 0;
-	params->protocol = NULL;
-	params->operation = UNDEFINED;
-	params->path = NULL;
-	params->table = NULL;
-	params->policy = NULL;
-	params->args = args;
-	
-	return params;
-}
-
-api_result check_parameters(rule_params* params)
-{
-	if(!params)
-		return INVALID;
-
-	switch(params->args)
-	{
-		case ARGS_IP:
-			return params->ip ? OK : INVALID_IP;
-		case ARGS_IP_PORT:
-			if(!params->ip) return INVALID_IP;
-			if(!params->port[0]) return INVALID_PORT;
-			if(!params->protocol) return INVALID_PROTOCOL;
-			return OK;
-		case ARGS_IP_PORT_RANGE:
-			if(!params->ip) return INVALID_IP;
-			if(!params->port[0]) return INVALID_PORT;
-			if(!params->port[1]) return INVALID_PORT;
-			if(params->port[1] < params->port[0]) return INVALID_PORT_RANGE;
-			if(!params->protocol) return INVALID_PROTOCOL;
-			return OK;
-		case ARGS_IP_SERVICE:
-			if(!params->ip) return INVALID_IP;
-			if(!params->service) return INVALID_SERVICE;
-			if(!params->port[0]) return INVALID_SERVICE;
-			if(!params->protocol) return INVALID_PROTOCOL;
-			return OK;
-		case ARGS_PORT:
-			if(!params->port[0]) return INVALID_PORT;
-			if(!params->protocol) return INVALID_PROTOCOL;
-			return OK;
-		case ARGS_PORT_RANGE:
-			if(!params->port[0]) return INVALID_PORT;
-			if(!params->port[1]) return INVALID_PORT;
-			if(params->port[1] < params->port[0]) return INVALID_PORT_RANGE;
-			if(!params->protocol) return INVALID_PROTOCOL;
-			return OK;
-		case ARGS_SERVICE:
-			if(!params->service) return INVALID_SERVICE;
-			if(!params->protocol) return INVALID_PROTOCOL;
-			return OK;
-		case ARGS_SAVE:
-		case ARGS_LOAD:
-			return params->path ? OK : INVALID_FILE_PATH;
-		case ARGS_CLEAR:
-			return OK;
-		case ARGS_POLICY_IN:
-		case ARGS_POLICY_OUT:
-			return params->policy ? OK : INVALID_POLICY;
-		default:
-			return INVALID;
-	}
-}
-
-DBusMessage* signal_from_rule_params(rule_params* params)
-{
-	DBusMessage* signal = NULL;
-	gchar *port_str = port_to_str(params);
-	const gchar *empty = EMPTY_STR;
-	const gchar *op = OP_STR[params->operation];
-
-	switch(params->args)
-	{
-		case ARGS_IP:
-			signal = sailfish_iptables_dbus_signal(
-				SAILFISH_IPTABLES_SIGNAL_RULE,
-				DBUS_TYPE_STRING,	&(params->ip),
-				DBUS_TYPE_STRING,	&empty,
-				DBUS_TYPE_STRING,	&op,
-				DBUS_TYPE_INVALID);
-			break;
-		case ARGS_IP_PORT:
-		case ARGS_IP_PORT_RANGE:
-		case ARGS_IP_SERVICE:
-			signal = sailfish_iptables_dbus_signal(
-				SAILFISH_IPTABLES_SIGNAL_RULE,
-				DBUS_TYPE_STRING,	&(params->ip),
-				DBUS_TYPE_STRING,	&port_str,
-				DBUS_TYPE_STRING,	&op,
-				DBUS_TYPE_INVALID);
-			break;
-		case ARGS_PORT:
-		case ARGS_PORT_RANGE:
-		case ARGS_SERVICE:
-			signal = sailfish_iptables_dbus_signal(
-				SAILFISH_IPTABLES_SIGNAL_RULE,
-				DBUS_TYPE_STRING,	&empty,
-				DBUS_TYPE_STRING,	&port_str,
-				DBUS_TYPE_STRING,	&op,
-				DBUS_TYPE_INVALID);
-			break;
-		case ARGS_SAVE:
-			signal = sailfish_iptables_dbus_signal(
-				SAILFISH_IPTABLES_SIGNAL_SAVE,
-				DBUS_TYPE_INVALID);
-			break;
-		case ARGS_LOAD:
-			signal = sailfish_iptables_dbus_signal(
-				SAILFISH_IPTABLES_SIGNAL_LOAD,
-				DBUS_TYPE_INVALID);
-		case ARGS_CLEAR:
-			signal = sailfish_iptables_dbus_signal(
-				SAILFISH_IPTABLES_SIGNAL_CLEAR,
-				DBUS_TYPE_INVALID);
-			break;
-		case ARGS_POLICY_IN:
-			signal = sailfish_iptables_dbus_signal(
-				SAILFISH_IPTABLES_SIGNAL_POLICY,
-				DBUS_TYPE_STRING,	IPTABLES_CHAIN_INPUT,
-				DBUS_TYPE_STRING,	&(params->policy),
-				DBUS_TYPE_INVALID);
-			break;
-		case ARGS_POLICY_OUT:
-			signal = sailfish_iptables_dbus_signal(
-				SAILFISH_IPTABLES_SIGNAL_POLICY,
-				DBUS_TYPE_STRING,	IPTABLES_CHAIN_OUTPUT,
-				DBUS_TYPE_STRING,	&(params->policy),
-				DBUS_TYPE_INVALID);
-			break;
-	}
-	g_free(port_str);
-	return signal;
-}
-
-rule_params* get_parameters_from_message(DBusMessage* message, rule_args args)
-{
-	rule_params *params = rule_params_new(args);
-	DBusError* error = NULL;
-	
-	gchar *ip = NULL, *service = NULL, *protocol = NULL, *port_str = NULL;
-	gchar *path = NULL, *table = NULL, *policy = NULL, *operation = NULL;
-	dbus_uint16_t port[2] = {0};
-	rule_operation op = UNDEFINED;
-	gint index = 0;
-	
-	gboolean rval = false;
-	
-	switch(params->args)
-	{
-		case ARGS_IP:
-			rval = dbus_message_get_args(message, error,
-						DBUS_TYPE_STRING, &ip,
-						DBUS_TYPE_STRING, &operation,
-						DBUS_TYPE_INVALID);
-			break;
-		case ARGS_IP_PORT:
-			rval = dbus_message_get_args(message, error,
-						DBUS_TYPE_STRING, &ip,
-						DBUS_TYPE_UINT16, &port,
-						DBUS_TYPE_STRING, &protocol,
-						DBUS_TYPE_STRING, &operation,
-						DBUS_TYPE_INVALID);
-			break;
-		case ARGS_IP_PORT_RANGE:
-			rval = dbus_message_get_args(message, error,
-						DBUS_TYPE_STRING, &ip,
-						DBUS_TYPE_STRING, &port_str,
-						DBUS_TYPE_STRING, &protocol,
-						DBUS_TYPE_STRING, &operation,
-						DBUS_TYPE_INVALID);
-			break;
-		case ARGS_IP_SERVICE:
-			rval = dbus_message_get_args(message, error,
-						DBUS_TYPE_STRING, &ip,
-						DBUS_TYPE_STRING, &service,
-						DBUS_TYPE_STRING, &protocol,
-						DBUS_TYPE_STRING, &operation,
-						DBUS_TYPE_INVALID);
-			break;
-		case ARGS_PORT:
-			rval = dbus_message_get_args(message, error,
-						DBUS_TYPE_UINT16, &port,
-						DBUS_TYPE_STRING, &protocol,			
-						DBUS_TYPE_STRING, &operation,
-						DBUS_TYPE_INVALID);
-			break;
-		case ARGS_PORT_RANGE:
-			rval = dbus_message_get_args(message, error,
-						DBUS_TYPE_STRING, &port_str,
-						DBUS_TYPE_STRING, &protocol,
-						DBUS_TYPE_STRING, &operation,
-						DBUS_TYPE_INVALID);
-			break;
-		case ARGS_SERVICE:
-			rval = dbus_message_get_args(message, error,
-						DBUS_TYPE_STRING, &service,
-						DBUS_TYPE_STRING, &protocol,
-						DBUS_TYPE_STRING, &operation,
-						DBUS_TYPE_INVALID);
-			break;
-		case ARGS_SAVE:
-		case ARGS_LOAD:
-			rval = dbus_message_get_args(message, error,
-						DBUS_TYPE_STRING, &path,
-						DBUS_TYPE_INVALID);
-			break;
-		case ARGS_CLEAR:
-			break;
-		case ARGS_POLICY_IN:
-		case ARGS_POLICY_OUT:
-			rval = dbus_message_get_args(message, error,
-						DBUS_TYPE_STRING, &policy,
-						DBUS_TYPE_INVALID);
-			
-			break;
-	}
-	
-	if(error)
-	{
-		DBG("Error, %s %s",
-			!rval ? "Could not get args from dbus message" : "", error->message);
-		rule_params_free(params);
-		dbus_error_free(error);
-		return NULL;
-	}
-	
-	if(ip && g_utf8_validate(ip,-1,NULL) && validate_ip_address(IPV4,ip))
-	{
-		if(negated_ip_address(ip))
-		{
-			params->ip = format_ip(IPV4,&(ip[1]));
-			params->ip_negate = true;
-		}
-		else
-			params->ip = format_ip(IPV4,ip);
-	}
-	
-	// Protocol defined
-	if(protocol && g_utf8_validate(protocol, -1, NULL))
-	{
-		gchar* protocol_lowercase = g_utf8_strdown(protocol, -1);
-
-		if(validate_protocol(protocol_lowercase))
-			params->protocol = protocol_lowercase;
-	}
-	
-	// Service defined
-	if(service && g_utf8_validate(service,-1,NULL))
-	{
-		gchar* service_lowercase = g_utf8_strdown(service,-1);
-		
-		// Check if the service with given name can be found, port and
-		// protocol can be retrieved then also
-		if((params->port[0] = validate_service_name(service_lowercase)))
-		{
-			params->service	= service_lowercase;
-			
-			if(!params->protocol)
-				params->protocol = get_protocol_for_service(params->service);
-		}
-	}
-	
-	// Port in string format
-	if(port_str && *port_str)
-	{
-		gchar **tokens = get_port_range_tokens(port_str);
-		
-		if(tokens)
-		{
-			for(index = 0; index < 2 && tokens[index] ; index++)
-				port[index] = (guint16)g_ascii_strtoull(tokens[index],NULL,10);
-		
-			// No second port was found, treat as ARGS_PORT/ARGS_IP_PORT
-			if(!port[1])
-			{
-				if(params->args == ARGS_IP_PORT_RANGE)
-					params->args = ARGS_IP_PORT;
-				else if(params->args == ARGS_PORT_RANGE)
-					params->args = ARGS_PORT;
-			}
-		}
-		
-		g_strfreev(tokens);
-	}
-	
-	// Check both ports
-	for(index = 0; index < 2 ; index++)
-	{
-		if(port[index] && validate_port(port[index]))
-		{
-			params->port[index] = port[index];
-
-			if(!params->protocol)
-				params->protocol = get_protocol_for_port(params->port[index]);
-		}
-	}
-	
-	// Operation defined
-	if(operation && *operation)
-	{
-		if((op = validate_operation(operation)) != UNDEFINED)
-			params->operation = op;
-	}
-	// No operation defined, defaults to ADD
-	else
-		params->operation = ADD;
-	
-	//
-	if(path && g_utf8_validate(path,-1,NULL) && validate_path(path))
-		params->path = g_strdup(path);
-	
-	if(table && *table && g_utf8_validate(table,-1,NULL))
-		params->table = g_utf8_strdown(table,-1);
-	
-	if(policy && g_utf8_validate(policy,-1,NULL))
-	{
-		gchar *policy_uppercase = g_utf8_strup(policy,-1);
-		if(validate_policy(policy_uppercase))
-			params->policy = policy_uppercase;
-	}
-		
-	return params;
-}
 
 static api_result clear_firewall(rule_params* params)
 {
